# backend/dist/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ZXC-BLOG API",
    "version": "1.0.0",
    "description": "API documentation for ZXC-BLOG"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login a user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/google": {
      "get": {
        "summary": "Authenticate with Google",
        "tags": [
          "Auth"
        ],
        "responses": {
          "302": {
            "description": "Redirect to Google for authentication"
          }
        }
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "summary": "Google authentication callback",
        "tags": [
          "Auth"
        ],
        "responses": {
          "302": {
            "description": "Redirect to the frontend application"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh the access token",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Successfully refreshed the token"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout a user",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          }
        }
      }
    },
    "/api/posts/{postId}/comments": {
      "get": {
        "summary": "Retrieve a list of comments for a post",
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of comments"
          }
        }
      },
      "post": {
        "summary": "Create a new comment",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a comment"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/comments/{id}": {
      "put": {
        "summary": "Update a comment",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateComment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated a comment"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Comment not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a comment",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted a comment"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Comment not found"
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "summary": "Retrieve a list of posts",
        "tags": [
          "Posts"
        ],
        "responses": {
          "200": {
            "description": "A list of posts"
          }
        }
      },
      "post": {
        "summary": "Create a new post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a post"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "summary": "Retrieve a single post",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single post"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "put": {
        "summary": "Update a post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated a post"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted a post"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
    "/api/posts/{id}/like": {
      "post": {
        "summary": "Like a post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully liked a post"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "delete": {
        "summary": "Unlike a post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully unliked a post"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Register": {
        "type": "object",
        "required": [
          "email",
          "password",
          "username"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "Login": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "CreateComment": {
        "type": "object",
        "required": [
          "content",
          "authorId",
          "postId"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "authorId": {
            "type": "integer"
          },
          "postId": {
            "type": "integer"
          },
          "parentId": {
            "type": "integer"
          }
        }
      },
      "UpdateComment": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "CreatePost": {
        "type": "object",
        "required": [
          "title",
          "content",
          "authorId"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "authorId": {
            "type": "integer"
          },
          "published": {
            "type": "boolean"
          }
        }
      },
      "UpdatePost": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "published": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication and authorization"
    },
    {
      "name": "Comments",
      "description": "Comment management"
    },
    {
      "name": "Posts",
      "description": "Post management"
    }
  ]
}